Hack_A_Loud_VillainArc_Day2-main/MedIQ_Advisor/EmotionDetect/app1.py
import time
import numpy as np
import cv2
import streamlit as st
from tensorflow import keras
from keras.models import model_from_json
from keras.preprocessing.image import img_to_array
from streamlit_webrtc import webrtc_streamer, VideoTransformerBase, WebRtcMode

# load model
emotion_dict = {0: 'angry', 1: 'happy', 2: 'neutral', 3: 'sad', 4: 'surprise'}
json_file = open('emotion_model1.json', 'r')
loaded_model_json = json_file.read()
json_file.close()
classifier = model_from_json(loaded_model_json)
classifier.load_weights("emotion_model1.h5")

# load face
try:
    face_cascade = cv2.CascadeClassifier('haarcascade_frontalface_default.xml')
except Exception:
    st.write("Error loading cascade classifiers")

class VideoTransformer(VideoTransformerBase):
    def __init__(self):
        self.frame_count = 0
        self.start_time = None
        self.duration = 5  # seconds

    def transform(self, frame):
        if self.start_time is None:
            self.start_time = time.time()

        current_time = time.time()
        elapsed_time = current_time - self.start_time

        img = frame.to_ndarray(format="bgr24")
        img_gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
        faces = face_cascade.detectMultiScale(img_gray, scaleFactor=1.3, minNeighbors=5)

        for (x, y, w, h) in faces:
            cv2.rectangle(img=img, pt1=(x, y), pt2=(x + w, y + h), color=(255, 0, 0), thickness=2)
            roi_gray = img_gray[y:y + h, x:x + w]
            roi_gray = cv2.resize(roi_gray, (48, 48), interpolation=cv2.INTER_AREA)

            if np.sum([roi_gray]) != 0:
                roi = roi_gray.astype('float') / 255.0
                roi = img_to_array(roi)
                roi = np.expand_dims(roi, axis=0)
                prediction = classifier.predict(roi)[0]
                maxindex = int(np.argmax(prediction))
                finalout = emotion_dict[maxindex]
                output = str(finalout)

            label_position = (x, y)
            cv2.putText(img, output, label_position, cv2.FONT_HERSHEY_SIMPLEX, 1, (0, 255, 0), 2)

        self.frame_count += 1

        if elapsed_time >= self.duration or self.frame_count >= 30:
            # Stop the webcam after 5 seconds or 30 frames
            webrtc_streamer.stop()

        return img

def main():
    st.title("Real Time Face Emotion Detection Application")
    activities = ["Webcam Face Detection"]
    choice = st.sidebar.selectbox("Select Activity", activities)
    
    if choice == "Webcam Face Detection":
        st.header("Webcam Live Feed")
        st.warning("Camera will automatically stop after 5 seconds. Click on 'Stop' to end the stream.")
        st.write("Click on start to use webcam and detect your face emotion")
        webrtc_streamer(key="example", video_transformer_factory=VideoTransformer, rtc_configuration={"video": True, "audio": False}, mode=WebRtcMode.SENDRECV)

if __name__ == "__main__":
    main()
